+++ trieste/models/__init__.py
@@ -23,6 +23,7 @@ from .interfaces import (
     FastUpdateModel,
     ModelStack,
     ProbabilisticModel,
+    ProbabilisticModelType,
     ReparametrizationSampler,
     TrainableModelStack,
     TrainableProbabilisticModel,
