+++ trieste/utils/__init__.py
@@ -12,35 +12,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """ This package contains library utilities. """
-from typing import Callable, Mapping, TypeVar
-
-from .misc import DEFAULTS, Err, Ok, Result, ResultType, jit, shapes_equal, to_numpy
-
-K = TypeVar("K")
-""" An unbound type variable. """
-
-U = TypeVar("U")
-""" An unbound type variable. """
-
-V = TypeVar("V")
-""" An unbound type variable. """
-
-
-def map_values(f: Callable[[U], V], mapping: Mapping[K, U]) -> Mapping[K, V]:
-    """
-    Apply ``f`` to each value in ``mapping`` and return the result. If ``f`` does not modify its
-    argument, :func:`map_values` does not modify ``mapping``. For example:
-
-    >>> import math
-    >>> squares = {'a': 1, 'b': 4, 'c': 9}
-    >>> map_values(math.sqrt, squares)['b']
-    2.0
-    >>> squares
-    {'a': 1, 'b': 4, 'c': 9}
-
-    :param f: The function to apply to the values in ``mapping``.
-    :param mapping: A mapping.
-    :return: A new mapping, whose keys are the same as ``mapping``, and values are the result of
-        applying ``f`` to each value in ``mapping``.
-    """
-    return {k: f(u) for k, u in mapping.items()}
+from .misc import (
+    DEFAULTS,
+    Err,
+    K,
+    Ok,
+    Result,
+    ResultType,
+    U,
+    V,
+    jit,
+    map_values,
+    shapes_equal,
+    to_numpy,
+)
