+++ trieste/utils/misc.py
@@ -14,7 +14,7 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import Any, Callable, Generic, NoReturn, TypeVar
+from typing import Any, Callable, Generic, Mapping, NoReturn, TypeVar
 
 import numpy as np
 import tensorflow as tf
@@ -180,3 +180,33 @@ class DEFAULTS:
     The default jitter, typically used to stabilise computations near singular points, such as in
     Cholesky decomposition.
     """
+
+
+K = TypeVar("K")
+""" An unbound type variable. """
+
+U = TypeVar("U")
+""" An unbound type variable. """
+
+V = TypeVar("V")
+""" An unbound type variable. """
+
+
+def map_values(f: Callable[[U], V], mapping: Mapping[K, U]) -> Mapping[K, V]:
+    """
+    Apply ``f`` to each value in ``mapping`` and return the result. If ``f`` does not modify its
+    argument, :func:`map_values` does not modify ``mapping``. For example:
+
+    >>> import math
+    >>> squares = {'a': 1, 'b': 4, 'c': 9}
+    >>> map_values(math.sqrt, squares)['b']
+    2.0
+    >>> squares
+    {'a': 1, 'b': 4, 'c': 9}
+
+    :param f: The function to apply to the values in ``mapping``.
+    :param mapping: A mapping.
+    :return: A new mapping, whose keys are the same as ``mapping``, and values are the result of
+        applying ``f`` to each value in ``mapping``.
+    """
+    return {k: f(u) for k, u in mapping.items()}
